Seperated the composables and data into their own packages
