+++ app/src/main/java/com/example/sightreadingapp/composables/LeaderboardScreen.kt
@@ -1,0 +1,75 @@
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Button
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.LocalContentColor
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.example.sightreadingapp.LeaderboardEntry
+import com.example.sightreadingapp.UserProfile
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun LeaderboardScreen(userProfile: UserProfile?, onRestartClicked: () -> Unit) {
+    // Dummy leaderboard entries.
+    val dummyEntries = listOf(
+        LeaderboardEntry("Alice", 50),
+        LeaderboardEntry("Bob", 30),
+        LeaderboardEntry("Charlie", 20),
+        LeaderboardEntry("Dave", 10)
+    )
+    // Combine dummy entries with the current user's profile.
+    val userEntry = userProfile?.let { LeaderboardEntry(it.name, it.score) }
+    val allEntries = if (userEntry != null) dummyEntries + userEntry else dummyEntries
+    val sortedEntries = allEntries.sortedByDescending { it.score }
+
+    Scaffold(
+        topBar = { TopAppBar(title = { Text("Leaderboard") }) }
+    ) { paddingValues ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(paddingValues)
+                .padding(16.dp)
+        ) {
+            Text("Leaderboard", style = MaterialTheme.typography.titleLarge)
+            Spacer(modifier = Modifier.height(16.dp))
+            sortedEntries.forEachIndexed { index, entry ->
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Text("${index + 1}.", modifier = Modifier.width(30.dp))
+                    Text(
+                        entry.name,
+                        modifier = Modifier.weight(1f),
+                        color = if (userProfile != null && entry.name == userProfile.name)
+                            MaterialTheme.colorScheme.primary
+                        else LocalContentColor.current
+                    )
+                    Text(entry.score.toString())
+                }
+                Spacer(modifier = Modifier.height(8.dp))
+            }
+            Spacer(modifier = Modifier.height(24.dp))
+            Button(
+                onClick = onRestartClicked,
+                modifier = Modifier.align(Alignment.CenterHorizontally)
+            ) {
+                Text("Restart")
+            }
+        }
+    }
+}
