+++ app/src/main/java/com/example/sightreadingapp/data/Accidentals.kt
@@ -1,0 +1,7 @@
+package com.example.sightreadingapp.data
+
+enum class Accidentals(val accident: String){
+    NONE(""),
+    FLAT("♭"),
+    SHARP("♯")
+}
