+++ app/src/main/java/com/example/sightreadingapp/data/NoteOptions.kt
@@ -1,0 +1,11 @@
+package com.example.sightreadingapp.data
+
+enum class NoteOptions(val note: String) {
+    A("A"),
+    B("B"),
+    C("C"),
+    D("D"),
+    E("E"),
+    F("F"),
+    G("G");
+}
