+++ app/src/main/java/com/example/sightreadingapp/data/Questions.kt
@@ -1,0 +1,9 @@
+package com.example.sightreadingapp.data
+
+data class Question(
+    val id: Int,
+    val noteResource: Int,  // Drawable resource for the Treble Clef note.
+    val options: List<String>,
+    val correctAnswer: String
+
+)
